image-colorizer-app/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.1
numpy==1.26.0
